Renamed org.scalactic.opaques package as org.scalactic.opaquetypes.
