+++ dotty/scalactic/src/main/scala/org/scalactic/opaquetypes/End.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalactic.opaques
+package org.scalactic.opaquetypes
 
 /**
   * Object that can be used as an endpoint for <code>NonEmptyList</code> construction expressions
