+++ jvm/scalactic/src/main/scala/org/scalactic/anyvals/NonEmptyString.scala
@@ -151,7 +151,7 @@ import org.scalactic.Resources
   * </pre>
   *
   */
-//DOTTY-ONLY @deprecated("Please use org.scalactic.opaques.NonEmptyString.", "3.3.0")  
+//DOTTY-ONLY @deprecated("Please use org.scalactic.opaquetypes.NonEmptyString.", "3.3.0")  
 final class NonEmptyString private (val theString: String) extends AnyVal {
 
   /**
@@ -1533,7 +1533,7 @@ final class NonEmptyString private (val theString: String) extends AnyVal {
 /**
   * Companion object for class <code>NonEmptyString</code>.
   */
-//DOTTY-ONLY @deprecated("Please use org.scalactic.opaques.NonEmptyString.", "3.3.0")  
+//DOTTY-ONLY @deprecated("Please use org.scalactic.opaquetypes.NonEmptyString.", "3.3.0")  
 object NonEmptyString {
 
   /**
